Improved detection of which player is "self"
Previously, "self" was assigned to whatever player's name appeared lowest on the screen (that's the most obvious way to determine it from the UI), but this led to rare misinterpretations, since the lowest name can sometimes be another player (when clicking player names to rotate them or when Possession is used).

Now the timers begin with that assumption, but they don't confirm it until the "waiting for (player name)" prompt has been detected for all players except one. If we detect we had the wrong player identified as "self", we convert all time we had previously assigned to them to the new player.